ml-service/requirements-registration.txt
# Requirements for model registration script
azure-ai-ml>=1.11.0
azure-identity>=1.12.0
transformers>=4.35.0
torch>=2.0.0
mlflow>=2.8.0
